+++ src/routes/(main)/+layout.svelte
@@ -76,6 +76,32 @@
 	</button>
 </div> -->
 
+<!-- <div
+	class="relative w-full bg-black text-white text-sm leading-sm px-10 md:px-8 py-6 font-light md:flex gap-3 justify-center"
+>
+	<div>
+		We're updating our Facilities page and your feedback helps! Take our
+		<a
+			href="https://docs.google.com/forms/d/e/1FAIpQLSfhGRg43hUUV4d229e44GfLFtEmS_vsJ8FaBOoS5MdHk-VNFw/viewform?usp=sharing&ouid=105526448285097294979"
+			target="_blank"
+			class="underline text-white"
+		>
+			survey
+		</a>.
+	</div>
+</div> -->
+
+<div
+	class="relative w-full bg-black text-white text-sm leading-sm px-10 md:px-8 py-6 font-light md:flex gap-3 justify-center"
+>
+	<div>
+		⚡️ Sign up for Open Electricity updates - direct to your inbox 📬.
+		<a href="https://openelectricity.org.au/newsletter" class="underline text-white"
+			>Subscribe now</a
+		>.
+	</div>
+</div>
+
 <!-- {#if feedbackButtonPosition}
 		<div
 			class="absolute z-50 bg-white border border-warm-grey rounded-lg px-8 py-6 shadow-lg"