+++ src/routes/(main)/studio/tracker-grid/+page.svelte
@@ -1,0 +1,27 @@
+<script>
+	import PageHeaderSimple from '$lib/components/PageHeaderSimple.svelte';
+	import Meta from '$lib/components/Meta.svelte';
+</script>
+
+<Meta
+	title="TrackerGrid"
+	description="TrackerGrid - A new data visualization tool in The Studio."
+	image="/img/preview.jpg"
+/>
+
+<PageHeaderSimple>
+	<div slot="main-heading">
+		<h1 class="tracking-widest text-center">TrackerGrid</h1>
+	</div>
+	<div slot="sub-heading">
+		<p class="text-sm text-center w-full md:w-[800px] mx-auto">
+			TrackerGrid visualization tool.
+		</p>
+	</div>
+</PageHeaderSimple>
+
+<section class="md:container py-12">
+	<div class="px-6">
+		<p class="text-center">TrackerGrid content coming soon...</p>
+	</div>
+</section>